+++ src/core/timers.ts
