+++ src/core/transactions/invite-server-transaction.ts
@@ -1,9 +1,9 @@
 import { C as SIPConstants } from "../../Constants";
 import { Exceptions } from "../../Exceptions";
-import { Timers } from "../../Timers";
-import { Transport } from "../../Transport";
 
 import { IncomingRequestMessage } from "../messages";
+import { Timers } from "../timers";
+import { Transport } from "../transport";
 import { ServerTransaction } from "./server-transaction";
 import { TransactionState } from "./transaction-state";
 import { ServerTransactionUser } from "./transaction-user";
