+++ src/core/user-agent-core/user-agent-core-configuration.ts
@@ -1,7 +1,8 @@
 import { DigestAuthentication } from "../../DigestAuthentication";
 import { LoggerFactory } from "../../LoggerFactory";
-import { Transport } from "../../Transport";
 import { URI } from "../../URI";
+
+import { Transport } from "../transport";
 
 /**
  * User agent contact.
